requirements.txt
django==5.0.2
djangorestframework==3.14.0
django-cors-headers==4.3.1
channels==4.0.0
daphne==4.1.0
better-sqlite3-django==1.0.0
